Added the Property name that represents the x and y coordinate to the options.
A lot of datasets have the 'x' and 'y' values called something else.
For example many of the datasets I deal with use 'Lat' And 'Lon' to
represent their coordinates.  This name is now configurable as part of
the options.  If no value is specified 'x' and 'y' are used.