Fixed Panning over international dateline issue & Issue with points located at the same lat/long
The following 3 issues were fixed:

1.) If the map was panned over the international dateline the
cluster/flares would no longer be drawn, unless the map was panned back
to the original starting point.  To fix this the webExtent is normalized
in the _clusterData function.

2.) If other layer types are added to the map which do not have an info
template or attributes assigned to the graphics, an exception would be
thrown in the _infoWindowShow function.  To fix this the
infoWindow.features is checked for null/undefined.

3.) If there are multiple points with the same lat/long, the flares
would all show the same data.  To fix this an optional idPropertyName
parameter was added that can be used to match the flare graphic with the
correct singleData.